feat(practice): add Practice/Catalog top-tab switcher rendering the catalog in place
Add an Insight-Timer-style centered Practice | Catalog switcher to the dark
Practice player. The Catalog tab renders the practice catalog in place on the
umber ground; activating a practice flips back to the Practice tab with the
new practice live via a single explicit refresh (no push nav, no double-fetch).

- Extract the catalog list body into PracticeCatalogList (variant + onActivated
  props); the pushed Catalog RootStack route stays byte-behavior-identical for
  deep links.
- New PracticeCatalogSwitcher with tablist/tab a11y roles, onShowcase active
  underline, and touch-target-sized tabs.
- Lift engine status to the screen so the switcher hides while a session runs
  or is paused; the drawer's in-place catalog rows are likewise withheld
  mid-session so they cannot unmount the running ritual.
