+++ frontend/src/features/Practice/components/PracticeDrawer.tsx
@@ -18,9 +18,15 @@ const ICON_SIZE = 20;
 
 export interface PracticeDrawerProps {
   hasActivePractice: boolean;
-  stageNumber: number;
   practiceId?: number;
   onCustomize: () => void;
+  /** Flip the Practice player to its embedded Catalog tab (no push nav). */
+  onBrowseCatalog: () => void;
+  /**
+   * True while the ritual engine is running or paused. The in-place catalog
+   * rows unmount that engine, so they are withheld until the session ends.
+   */
+  sessionActive: boolean;
   onClose: () => void;
 }
 
@@ -35,19 +41,23 @@ interface DrawerRow {
  * Builds the ordered rows for the current state. The active state exposes the
  * full set (change/browse/customize/details/create); the empty state offers only
  * browse and create. "Practice details" appears only when a practiceId resolves.
+ * Catalog rows flip the player's embedded Catalog tab in place; details/create
+ * remain pushed routes. While a session is running or paused the in-place
+ * catalog rows are withheld — flipping the tab would unmount the live engine.
  */
 function usePracticeRows({
   hasActivePractice,
-  stageNumber,
   practiceId,
   onCustomize,
+  onBrowseCatalog,
+  sessionActive,
 }: Omit<PracticeDrawerProps, 'onClose'>): DrawerRow[] {
   const navigation = useNavigation<NativeStackNavigationProp<RootStackParamList>>();
   const browse: DrawerRow = {
     testID: 'practice-drawer-browse',
     label: 'Browse all practices',
     Icon: Compass,
-    run: () => navigation.navigate('Catalog', { stageNumber }),
+    run: onBrowseCatalog,
   };
   const create: DrawerRow = {
     testID: 'practice-drawer-create',
@@ -56,21 +66,20 @@ function usePracticeRows({
     run: () => navigation.navigate('CreatePractice'),
   };
   if (!hasActivePractice) return [browse, create];
-  const rows: DrawerRow[] = [
-    {
-      testID: 'practice-drawer-change',
-      label: 'Change practice',
-      Icon: RefreshCw,
-      run: () => navigation.navigate('Catalog', { stageNumber }),
-    },
-    browse,
-    {
-      testID: 'practice-drawer-customize',
-      label: 'Customize this practice',
-      Icon: SlidersHorizontal,
-      run: onCustomize,
-    },
-  ];
+  const change: DrawerRow = {
+    testID: 'practice-drawer-change',
+    label: 'Change practice',
+    Icon: RefreshCw,
+    run: onBrowseCatalog,
+  };
+  const customize: DrawerRow = {
+    testID: 'practice-drawer-customize',
+    label: 'Customize this practice',
+    Icon: SlidersHorizontal,
+    run: onCustomize,
+  };
+  // Withhold the tab-flip rows mid-session; keep the modal/push-based rows.
+  const rows: DrawerRow[] = sessionActive ? [customize] : [change, browse, customize];
   if (practiceId !== undefined) {
     rows.push({
       testID: 'practice-drawer-details',
